app/requirements.txt
# Application dependencies
fastapi
uvicorn
polars
duckdb
streamlit
altair
python-dotenv

# Testing dependencies
pytest
httpx